dot: tutorials my first widget: split the results into three files
+ use "clicked" instead of "released" for the connection.
